refactor(core): extract the plugin-skill registry to a leaf (breaks the last runtime cycle)
The remaining dependency cycle was closed by workflow-skill-drift.ts
importing getPluginSkills from the plugin loader. The plugin-skill map +
getPluginSkills/removePluginSkillsByPlugin/clearPluginSkills move to a
dependency-free leaf (@bakin/core/skills/plugin-skill-registry,
globalThis-backed like the hook singleton); the workflows plugin's skill
consumers and the reload pipeline import it from there, and plugin-registry
re-exports it for the existing surface. madge now reports 18 → 4 cycles:
both WS2-targeted runtime clusters (scripts/lib registry + workflow
registries) are gone; the 4 left are pre-existing type-only cycles (core
docs/plugin-types and the WS1 SDK split).

The workflow-registry boundary move (core importing plugins/workflows
internals) is deferred to WS6, where the workflows-plugin split happens —
they're intertwined. No behavior change.
