+++ src/core/plugin-host/reload-pipeline.ts
@@ -35,7 +35,8 @@
 import { existsSync } from 'fs'
 import { join } from 'path'
 import { createLogger } from '@/core/logger'
-import { pluginRegistry, removePluginSkillsByPlugin } from '@/lib/plugin-registry'
+import { pluginRegistry } from '@/lib/plugin-registry'
+import { removePluginSkillsByPlugin } from '@bakin/core/skills/plugin-skill-registry'
 import { getHookRegistry } from '@bakin/core/hooks/hook-registry-singleton'
 import { removeExecToolsByPlugin } from '../exec-tools/registry'
 import { unregisterPluginNodeTypes } from '../../../plugins/workflows/lib/node-type-registry'
